backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pytz
httpx
apscheduler
solapi
google-generativeai
Pillow
